core/requirements.txt
Django==2.2.7
djangorestframework==3.9.2
redis==3.4.1
uWSGI==2.0.18
